requirements.txt
Flask
Pillow
exifread
imagehash
